ExpandeView，LoadingDialog,添加流式布局 等

+++ baselib/src/main/java/com/jskingen/baselib/network/utils/FileRequestBody.java
@@ -77,7 +77,7 @@ public class FileRequestBody<T>  extends RequestBody {
                 //增加当前写入的字节数
                 bytesWritten += byteCount;
                 //回调
-                callback.onLoading(contentLength, bytesWritten);
+                callback.onLoading(contentLength, (int) bytesWritten);
             }
         };
     }
